+++ Libplanet.Node/UntypedBlock.cs
@@ -97,6 +97,9 @@ namespace Libplanet.Node
         /// <inheritdoc cref="IBlockMetadata.TxHash"/>
         public HashDigest<SHA256>? TxHash => _header.TxHash;
 
+        /// <inheritdoc cref="IBlockMetadata.PreviousMiner"/>
+        public Address? PreviousMiner => _header.PreviousMiner;
+
         /// <inheritdoc cref="IBlockMetadata.LastCommit"/>
         public BlockCommit? LastCommit => _header.LastCommit;