+++ Libplanet.Tests/Blocks/BlockMarshalerTest.cs
@@ -32,6 +32,7 @@ namespace Libplanet.Tests
         private static readonly byte[] StateRootHashKey = { 0x73 }; // 's'
         private static readonly byte[] SignatureKey = { 0x53 }; // 'S'
         private static readonly byte[] PreEvaluationHashKey = { 0x63 }; // 'c'
+        private static readonly byte[] PreviousMinerKey = { 0x4D }; // 'M'
         private static readonly byte[] LastCommitKey = { 0x43 }; // 'C'
 
         // Block fields:
@@ -81,6 +82,7 @@ namespace Libplanet.Tests
                     TimestampKey,
                     _fx.Next.Timestamp.ToString(TimestampFormat, CultureInfo.InvariantCulture))
                 .Add(PublicKeyKey, _fx.Next.PublicKey.Format(compress: true))
+                .Add(PreviousMinerKey, _fx.Next.PreviousMiner.Value.ByteArray)
                 .Add(LastCommitKey, _fx.Next.LastCommit.Value.ByteArray);
             var expectedNextHeader = _marshaledNextMetadata
                 .Add(PreEvaluationHashKey, _fx.Next.PreEvaluationHash.ByteArray)
@@ -99,6 +101,7 @@ namespace Libplanet.Tests
                     _fx.HasTx.Timestamp.ToString(TimestampFormat, CultureInfo.InvariantCulture))
                 .Add(PublicKeyKey, _fx.HasTx.PublicKey.Format(true))
                 .Add(TxHashKey, _fx.HasTx.TxHash.Value.ByteArray)
+                .Add(PreviousMinerKey, _fx.HasTx.PreviousMiner.Value.ByteArray)
                 .Add(LastCommitKey, _fx.HasTx.LastCommit.Value.ByteArray);
             var expectedHasTxHeader = _marshaledHasTxMetadata
                 .Add(PreEvaluationHashKey, _fx.HasTx.PreEvaluationHash.ByteArray)