+++ Libplanet/Blocks/Block.cs
@@ -116,6 +116,9 @@ namespace Libplanet.Blocks
         /// <inheritdoc cref="IBlockMetadata.TxHash"/>
         public HashDigest<SHA256>? TxHash => _preEvaluationBlock.TxHash;
 
+        /// <inheritdoc cref="IBlockMetadata.PreviousMiner"/>
+        public Address? PreviousMiner => _preEvaluationBlock.PreviousMiner;
+
         /// <inheritdoc cref="IBlockMetadata.LastCommit"/>
         public BlockCommit? LastCommit => _preEvaluationBlock.LastCommit;