+++ Libplanet/Blocks/IBlockMetadata.cs
@@ -54,6 +54,11 @@ namespace Libplanet.Blocks
         HashDigest<SHA256>? TxHash { get; }
 
         /// <summary>
+        /// The address of the miner on previous block.
+        /// </summary>
+        Address? PreviousMiner { get; }
+
+        /// <summary>
         /// The <see cref="BlockCommit"/> about previous block's vote information.
         /// </summary>
         BlockCommit? LastCommit { get; }