+++ Libplanet/Blocks/InvalidBlockPreviousMinerException.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace Libplanet.Blocks
+{
+    [Serializable]
+    public class InvalidBlockPreviousMinerException : InvalidBlockException
+    {
+        public InvalidBlockPreviousMinerException(string message)
+            : base(message)
+        {
+        }
+    }
+}