+++ Libplanet/Blocks/PreEvaluationBlockHeader.cs
@@ -220,6 +220,8 @@ namespace Libplanet.Blocks
         /// <inheritdoc cref="IBlockMetadata.TxHash"/>
         public HashDigest<SHA256>? TxHash => Metadata.TxHash;
 
+        public Address? PreviousMiner => Metadata.PreviousMiner;
+
         public BlockCommit? LastCommit => Metadata.LastCommit;
 
         /// <inheritdoc cref="IPreEvaluationBlockHeader.PreEvaluationHash"/>