+++ Libplanet/Store/BlockDigest.cs
@@ -83,6 +83,9 @@ namespace Libplanet.Store
         /// <inheritdoc cref="IBlockMetadata.TxHash"/>
         public HashDigest<SHA256>? TxHash => _metadata.TxHash;
 
+        /// <inheritdoc cref="IBlockMetadata.PreviousMiner"/>
+        public Address? PreviousMiner => _metadata.PreviousMiner;
+
         /// <inheritdoc cref="IBlockMetadata.LastCommit"/>
         public BlockCommit? LastCommit => _metadata.LastCommit;